requirements.txt
asgiref==3.8.1
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
idna==3.7
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
packaging==24.0
pillow==10.3.0
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pywin32==306
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
